docker-entrypoint: use a copy of users .gitconfig instead of the volume
The hosts .gitconfig is mounted from the host to the container as a volume.
It's possible to have access problems if multiple containers or the host
is trying to access the file. This can manifest with the following
error when Yocto tries to modify .gitconfig:

    | error: could not lock config file /home/vari/.gitconfig: Permission denied
    | WARNING: /workdir/build_xwayland/tmp/work/armv8a-fslc-linux/onnxruntime/1.10.0-r0/temp/run.do_compile.4081559:159 exit 255 from 'git config --global --add safe.directory /workdir/build_xwayland/tmp/work/armv8a-fslc-linux/onnxruntime/1.10.0-r0/build/pybind11/src/pybind11'
    | WARNING: Backtrace (BB generated script):
    |       #1: do_compile, /workdir/build_xwayland/tmp/work/armv8a-fslc-linux/onnxruntime/1.10.0-r0/temp/run.do_compile.4081559, line 159
    |       #2: main, /workdir/build_xwayland/tmp/work/armv8a-fslc-linux/onnxruntime/1.10.0-r0/temp/run.do_compile.4081559, line 184

Fix this by using a copy of .gitconfig inside the container. This also
adds the benefit of not modifying the .gitconfig of the host.

Signed-off-by: Nate Drude <nate.d@variscite.com>