Dockerfile_14.04: Drop git-lfs package
The image previously installed git-lfs using the official
packagecloud.io/github/git-lfs installer script. As of November 2025, the
script completes on ubuntu:14.04 but `apt install git-lfs` fails with
"Unable to locate package git-lfs".

Because this install path is no longer reliable, git-lfs is removed
from the image. If git-lfs needs to be restored in the future, it
should be installed from the upstream binary tarball rather than
via packagecloud.

Signed-off-by: Nate Drude <nate.d@variscite.com>